requirements.txt
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
scikit-learn==1.5.2
scipy==1.13.1
torch==2.0.0
torchvision==0.15.1

